fix: Cache busting for entrypoint.
The entrypoint to the app was always named `index.js`, and we set long
cache expiration times on `.js` files. To fix this, we can use esbuild's
option to inject a hash into the name, but we have to make sure we
replace the script reference in our root HTML file with the unique name.