chore: Prepare for deploy to S3.
Set up the ability to push files to S3. This allows static files to
persist longer than the current deployment, which is required for the
case where a user loads the page, a new version is pushed, then the user
navigates to a section of the page that requires a chunk load.