feat: Add landing page
It seems like a better user experience to see a landing page instead of
being instantly asked to log in. We still preserve the behavior of
authenticated users being able to visit the app at the root domain by
redirecting authenticated users from the landing page to the app unless
a specific query parameter is set.